fixed bug rollback api
